Refactor
Refactored to avoid ugly inheritances from core
